+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Traits description on automata cores
 - Base automata levels: netherite, starbound and creative
 - Villager automata levels: starbound and creative
+- Trashing turtle
 
 ### Changed